pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfy-cli"
license = {file = "LICENSE"}
version = "2024.0.0"
requires-python = ">= 3.8"
description = "A CLI tool for installing and using ComfyUI."
readme = "README.md"
keywords = ["comfyui", "stable diffusion"]

maintainers = [
  {name = "Yoland Yan", email = "yoland@drip.art"},
  {name = "James Kwon", email = "hongilkwon316@gmail.com"},
  {name = "Robin Huang", email = "robin@drip.art"}
]

dependencies = [
    "typer>=0.3.0",
    "GitPython",
    "rich",
    "requests",
    "pyyaml",
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"
]

[project.scripts]
comfy = "comfy.__main__:app"
comfy-cli = "comfy.__main__:app"
comfycli = "comfy.__main__:app"

[project.urls]
Repository = "https://github.com/drip-art/comfy-cli.git"



